+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 # [Unreleased]
 ### Fixed
 - Fixed potential panic if a benchmark took zero time.
+- cargo-criterion now calls `cargo metadata` to find the path to the target directory. This fixes
+  the location of the target directory in workspaces.
 
 ## [1.0.0-alpha3] - 2020.07-06
 ### Added